pyproject.toml
[tool.poetry]
name = "fitness-bot"
version = "0.1.0"
description = ""
authors = ["dmitriy <dmitriydmgora@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.6.0"
aiogram-dialog = "^2.1.0"
redis = "^5.0.4"
environs = "^11.0.0"
fluentogram = "^1.1.7"
nats-py = "^2.8.0"
ormsgpack = "^1.5.0"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
psycopg = {extras = ["binary"], version = "^3.2.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
